[dxvk] Expose a way to increment stat counters
In case the counters come from external sources.

+++ src/dxvk/dxvk_cmdlist.h
@@ -102,7 +102,7 @@ namespace dxvk {
      * \param [in] ctr The counter to increment
      * \param [in] val The value to add
      */
-    void addStatCtr(DxvkStatCounter ctr, uint32_t val) {
+    void addStatCtr(DxvkStatCounter ctr, uint64_t val) {
       m_statCounters.addCtr(ctr, val);
     }
